+++ app/apis/api/v1/event_log_api.rb
@@ -1,20 +1,18 @@
 module API
   module V1
     class EventLogAPI < Grape::API
-      content_type :json, 'application/json'
-      content_type :csv,  'application/csv'
+      helpers Helper::SharedParams
 
       before { authenticate_user! }
 
       desc(
         'Get logs of TodoList',
         tags: ['logs'],
-        success: Entity::V1::EventLog,
+        success: Entity::V1::PaginatedEventLog,
         is_array: true
       )
       params do
-        requires :start_date, type: Date, default: Date.today
-        requires :end_date, type: Date, default: Date.today
+        use :period, :pagination
       end
       get 'todo_list/:todo_lis_id/logs' do
         start_date = params[:start_date].beginning_of_day
@@ -23,9 +21,8 @@ module API
 
         todo_list = current_user.todo_lists.find(params[:todo_lis_id])
         event_logs = EventLog.where(log_tag: todo_list.log_tag)
-        error!('No data.', 204) unless event_logs.present?
 
-        present event_logs, with: Entity::V1::EventLog
+        paginate event_logs, with: Entity::V1::PaginatedEventLog
       end
     end
   end