+++ app/models/token.rb
@@ -1,7 +1,7 @@
 class Token < ApplicationRecord
   SCOPE_OPTIONS = %w[read:log write:log].freeze
 
-  belongs_to :user
+  belongs_to :user, optional: true
 
   def payload
     {